refactor: centralize cache clearing utility
